test(browser): move workspace stamp tests into a colocated tests module
The inline test block had grown past the module's production code, so the
tests now live in browser/manifest/workspace/tests.rs per the maintainability
rule; no production code changes.
